<feat> 제품 정보 공유 게시판 수정, 찜, 북마크 기능 구현

+++ src/main/java/com/project/compagnoserver/controller/ProductBoardController.java
@@ -1,8 +1,6 @@
 package com.project.compagnoserver.controller;
 
-import com.project.compagnoserver.domain.ProductBoard.ProductBoard;
-import com.project.compagnoserver.domain.ProductBoard.ProductBoardDTO;
-import com.project.compagnoserver.domain.ProductBoard.ProductBoardImage;
+import com.project.compagnoserver.domain.ProductBoard.*;
 import com.project.compagnoserver.service.ProductBoardService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -33,18 +31,20 @@ public class ProductBoardController {
     private ProductBoardService productBoard;
 
     @Value("${spring.servlet.multipart.location}")
-    private String uploadPath; // D:\\upload
+    private String uploadPath; // C:\\upload
 
     // 게시물 등록
     @PostMapping("/productBoard")
     public ResponseEntity<ProductBoard> create(ProductBoardDTO dto) throws IOException {
-
+        String saveName = null;
         // 메인 이미지 업로드
-        String fileName = dto.getProductMainFile().getOriginalFilename();
-        String uuid = UUID.randomUUID().toString();
-        String saveName = uploadPath + File.separator + "productBoardMainImage" + File.separator + uuid + "_" + fileName;
-        Path savePath = Paths.get(saveName);
-        dto.getProductMainFile().transferTo(savePath);
+        if(!dto.getProductMainFile().isEmpty()) {
+            String fileName = dto.getProductMainFile().getOriginalFilename();
+            String uuid = UUID.randomUUID().toString();
+            saveName = uploadPath + File.separator + "productBoardMainImage" + File.separator + uuid + "_" + fileName;
+            Path savePath = Paths.get(saveName);
+            dto.getProductMainFile().transferTo(savePath);
+        }
 
         // 게시판 작성
         ProductBoard vo = ProductBoard.builder()
@@ -62,16 +62,18 @@ public class ProductBoardController {
 
         // 이미지 업로드
         for (MultipartFile file : dto.getFiles()) {
-            ProductBoardImage imgVo = new ProductBoardImage();
-            fileName = file.getOriginalFilename();
-            uuid = UUID.randomUUID().toString();
-            saveName = uploadPath + File.separator + "productBoardImage" + File.separator + uuid + "_" + fileName;
-            savePath = Paths.get(saveName);
-            file.transferTo(savePath);
+            String fileName = file.getOriginalFilename();
+            if(!fileName.isEmpty()) {
+                ProductBoardImage imgVo = new ProductBoardImage();
+                String uuid = UUID.randomUUID().toString();
+                saveName = uploadPath + File.separator + "productBoardImage" + File.separator + uuid + "_" + fileName;
+                Path savePath = Paths.get(saveName);
+                file.transferTo(savePath);
 
-            imgVo.setProductBoard(result);
-            imgVo.setProductImage(saveName);
-            productBoard.createImage(imgVo);
+                imgVo.setProductBoard(result);
+                imgVo.setProductImage(saveName);
+                productBoard.createImage(imgVo);
+            }
         }
 
         return result!=null ?
@@ -86,13 +88,15 @@ public class ProductBoardController {
         
         // 메인 이미지 삭제
         ProductBoard prev = productBoard.viewBoard(code);
-        File file = new File(prev.getProductMainImage());
-        file.delete();
+        if(prev.getProductMainImage()!=null) {
+            File file = new File(prev.getProductMainImage());
+            file.delete();
+        }
 
         // 나머지 이미지 삭제
         List<ProductBoardImage> prevImage = productBoard.viewImage(code);
         for(ProductBoardImage image : prevImage) {
-             file = new File(image.getProductImage());
+             File file = new File(image.getProductImage());
              file.delete();
         }
 
@@ -113,7 +117,7 @@ public class ProductBoardController {
     }
 
     // 조회수
-    private void viewCountUp(int code, HttpServletRequest req, HttpServletResponse res) {
+    public void viewCountUp(int code, HttpServletRequest req, HttpServletResponse res) {
         Cookie[] cookies = Optional.ofNullable(req.getCookies()).orElseGet(() -> new Cookie[0]);
 
         Cookie cookie = Arrays.stream(cookies)
@@ -133,5 +137,85 @@ public class ProductBoardController {
         res.addCookie(cookie);
     }
 
+    // 게시판 수정
+    @PutMapping("/productBoard")
+    public ResponseEntity<ProductBoard> update(ProductBoardDTO dto) throws IOException {
 
+        // 기존 이미지 삭제
+        ProductBoard prev = productBoard.viewBoard(dto.getProductBoardCode());
+        if(prev.getProductMainImage()!=null) {
+            File file = new File(prev.getProductMainImage());
+            file.delete();
+        }
+
+        // 나머지 이미지 삭제
+        List<ProductBoardImage> prevImage = productBoard.viewImage(dto.getProductBoardCode());
+        for(ProductBoardImage image : prevImage) {
+            File file = new File(image.getProductImage());
+            file.delete();
+            productBoard.deleteImage(dto.getProductBoardCode());
+        }
+
+        // 메인 이미지 업로드
+        String saveName = null;
+        if(!dto.getProductMainFile().isEmpty()) {
+            String fileName = dto.getProductMainFile().getOriginalFilename();
+            String uuid = UUID.randomUUID().toString();
+            saveName = uploadPath + File.separator + "productBoardMainImage" + File.separator + uuid + "_" + fileName;
+            Path savePath = Paths.get(saveName);
+            dto.getProductMainFile().transferTo(savePath);
+        }
+
+        // 게시판 작성
+        ProductBoard vo = ProductBoard.builder()
+                .productBoardCode(dto.getProductBoardCode())
+                .productBoardTitle(dto.getProductBoardTitle())
+                .productMainImage(saveName)
+                .productName(dto.getProductName())
+                .productPrice(dto.getProductPrice())
+                .productCategory(dto.getProductCategory())
+                .productBoardGrade(dto.getProductBoardGrade())
+                .productBoardContent(dto.getProductBoardContent())
+                .userId(dto.getUserId())
+                .animalCategoryCode(dto.getAnimalCategoryCode())
+                .build();
+        ProductBoard result = productBoard.updateBoard(vo);
+
+        // 이미지 업로드
+        for (MultipartFile file : dto.getFiles()) {
+            String fileName = file.getOriginalFilename();
+            if(!fileName.isEmpty()) {
+                ProductBoardImage imgVo = new ProductBoardImage();
+                String uuid = UUID.randomUUID().toString();
+                saveName = uploadPath + File.separator + "productBoardImage" + File.separator + uuid + "_" + fileName;
+                Path savePath = Paths.get(saveName);
+                file.transferTo(savePath);
+
+                imgVo.setProductBoard(result);
+                imgVo.setProductImage(saveName);
+            }
+        }
+
+        return result!=null ?
+                ResponseEntity.status(HttpStatus.CREATED).body(result) :
+                ResponseEntity.badRequest().build();
+    }
+
+    // 게시판 추천, 추천 된 상태면 삭제
+    @PostMapping("/productBoard/recommend")
+    public ResponseEntity boardRecommend(@RequestBody ProductBoardRecommend vo) {
+        vo.setUserId("ghldud"); // 나중에 user 정보
+        productBoard.boardRecommend(vo);
+
+        return ResponseEntity.ok().build();
+    }
+
+    // 게시판 북마크, 북마크 된 상태면 삭제
+    @PostMapping("/productBoard/bookmark")
+    public ResponseEntity boardBookmark(@RequestBody ProductBoardBookmark vo) {
+        vo.setUserId("ghldud"); // 나중에 user 정보
+        productBoard.boardBookmark(vo);
+
+        return ResponseEntity.ok().build();
+    }
 }
